add spring security

+++ CourseProject/thds_backend/src/main/java/phearun/thds_backend/repository/IssueRepository.java
@@ -1,9 +1,36 @@
 package phearun.thds_backend.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
+import phearun.thds_backend.dto.IssueDetailDTO;
 import phearun.thds_backend.model.Issue;
 
+import java.util.List;
+
 @Repository
-public interface IssueRepository extends JpaRepository<Issue, Integer> {
+public interface IssueRepository extends JpaRepository<Issue, Long>{
+
+//,
+
+
+//    @Query(value = "select iss.iss_id, iss.iss_name from issue as iss inner join subcategory as sub on iss.sub_cate_id = sub.id",
+//            nativeQuery = true)
+//    List<Issue> findAllIssue();
+
+//    @Query(value = "SELECT iss.iss_name, sub.subcate_name FROM issue iss INNER JOIN subcategory sub on iss.sub_cate_id = sub.id", nativeQuery = true)
+//    List<IssueDetailDTO> findAllIssue();
+
+//    public interface UserRepository extends CrudRepository<UserEntity, Long> {
+//        @Query(value = "select new com.example.package.StatsDTO(count(type_id) userCount, typeId, modifiedAt from "
+//                + "UserCampaignObjective where campId = ?1 group by objectiveTypeId,modifiedAt")
+//
+//        List<StatsDTO> getStatsDTO(Long camp_id);
+//    }
+
+
+
+
+
 }
