+++ CourseProject/thds_backend/src/main/java/phearun/thds_backend/service/IssueService.java
@@ -1,6 +1,7 @@
 package phearun.thds_backend.service;
 
 
+import phearun.thds_backend.dto.IssueDetailDTO;
 import phearun.thds_backend.model.Issue;
 
 import java.util.List;
@@ -8,11 +9,13 @@ import java.util.List;
 public interface IssueService {
     Issue save(Issue entity);
 
-    Issue getById(Integer id);
+    Issue getById(Long id);
 
-    Issue update(Integer id, Issue Issue) ;
+    Issue update(Long id, Issue Issue) ;
 
     List<Issue> getIssues();
 
-    void delete(Integer id);
+   // List<IssueDetailDTO>findAllIssue();
+
+    void delete(Long id);
 }
